Minor changes * Update resources under docs * Get rid of warnings in code * Reformat classes * Simplify expressions

+++ src/utils/ShaderLoader.cpp
@@ -27,18 +27,16 @@ GLuint ShaderLoader::load(const char *vertex_path, const char *fragment_path) {
     std::cout << "Compiling vertex shader." << std::endl;
     glShaderSource(vertShader, 1, &vertShaderSrc, NULL);
     glCompileShader(vertShader);
-   // std::cout << "yeah" << glGetError() << std::endl;
 
     // Check vertex shader
     glGetShaderiv(vertShader, GL_COMPILE_STATUS, &result);
     glGetShaderiv(vertShader, GL_INFO_LOG_LENGTH, &logLength);
-    //std::cout << "yeah" << glGetError() << std::endl;
+
     if (logLength > 0) {
-        std::vector<char> vertShaderError(logLength);
+        std::vector<char> vertShaderError((unsigned long) logLength);
         glGetShaderInfoLog(vertShader, logLength, NULL, &vertShaderError[0]);
         std::cout << &vertShaderError[0] << std::endl;
     }
-
 
     // Compile fragment shader
     std::cout << "Compiling fragment shader." << std::endl;
@@ -50,7 +48,7 @@ GLuint ShaderLoader::load(const char *vertex_path, const char *fragment_path) {
     glGetShaderiv(fragShader, GL_COMPILE_STATUS, &result);
     glGetShaderiv(fragShader, GL_INFO_LOG_LENGTH, &logLength);
     if (logLength > 0) {
-        std::vector<char> fragShaderError(logLength);
+        std::vector<char> fragShaderError((unsigned long) logLength);
         glGetShaderInfoLog(fragShader, logLength, NULL, &fragShaderError[0]);
         std::cout << &fragShaderError[0] << std::endl;
     }
@@ -62,7 +60,7 @@ GLuint ShaderLoader::load(const char *vertex_path, const char *fragment_path) {
 
     glGetProgramiv(program, GL_LINK_STATUS, &result);
     glGetProgramiv(program, GL_INFO_LOG_LENGTH, &logLength);
-    std::vector<char> programError((logLength > 1) ? logLength : 1);
+    std::vector<char> programError((unsigned long) ((logLength > 1) ? logLength : 1));
     glGetProgramInfoLog(program, logLength, NULL, &programError[0]);
     std::cout << &programError[0] << std::endl;
 
@@ -72,7 +70,7 @@ GLuint ShaderLoader::load(const char *vertex_path, const char *fragment_path) {
     return program;
 }
 
-ShaderLoader::ShaderLoader(const ShaderLoader& orig) {
+ShaderLoader::ShaderLoader(const ShaderLoader &orig) {
 }
 
 ShaderLoader::~ShaderLoader() {
